chatbot/requirements.txt
colorama
google-generativeai
requests
flask
flask-cors
